+++ SBCK/__release.py
@@ -20,13 +20,12 @@
 version_major = 2
 version_minor = 0
 version_patch = 0
-version_extra = "a56"
+version_extra = "a57"
 version       = "{}.{}.{}{}".format(version_major,version_minor,version_patch,version_extra)
 
 name         = "SBCK"
 description  = "Statistical Bias Correction Kit"
 author       = "Yoann Robin"
-author_email = "yoann.robin.k@gmail.com"
-#license      = "GNU General Public License v3"
+author_email = "yoann.robin+dev.github@ikmail.com"
 license      = "GPL-3.0-or-later"